financial-organizer/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.7
pytest==7.4.4
httpx==0.26.0
pytest-cov==4.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pillow==10.2.0
pytesseract==0.3.10
pyzbar==0.1.9
pdf2image==1.17.0
plaid-python==15.5.0
aiofiles==23.2.1
pre-commit==3.6.0
email-validator==2.1.0 